feat(cardano): take the federation identity from the chain, and stop defaulting it (WI-069) (#38)
Y_federation and the recovery leaf's CSV delay build the Bitcoin treasury's Taproot address. Y_51
rotates each epoch with the DKG and arrives over the authenticated wire, but the federation leaf
does not rotate — every SPO rebuilds the tree locally, so a node holding a different value derives
a different scriptPubKey, signs a different BIP-341 sighash, and contributes a FROST share over a
message no other signer produced. Nothing errors: the address it derived is well-formed, it is
just not the one the BTC is in. Both were hand-copied from deployment notes into every operator's
config. They were already on chain the whole time — the treasury_info datum carries them at fields
#2/#3, heimdall writes them there at bootstrap, and treasury.ak enforces that copy for Update-Y —
so the fix reads what exists rather than publishing a second copy. The whole path hangs off the
one Config NFT: #12/#13 name the treasury_info policy and state NFT, its script hash gives the
address, and the snapshot heimdall already fetches to cross-check the identity trie carries the
datum. No new field, no extra query, no on-chain change.

The defaults are gone, which was the sharper half: y_fed_seed_hex defaulted to 0xFE x 32 and
federation_csv_blocks to 144, so an operator who set neither got no error, they got a confidently
derived wrong treasury. Both are Option, unset means "read the chain", and nowhere-configured is
fatal. federation_csv_blocks is now u16 end to end — it was a u32 that every consumer cast, so a
value above 65535 truncated into a different tree — and is validated where it enters. The identity
is resolved ONCE at startup and passed down, because the mover previously re-derived it from the
local seed and would have died on its first tick under the very config this change tells operators
to write. The seed keeps its one irreplaceable job, signing federation spends, so bootstrap-treasury,
federation-spend and the depositor demo still require it by name and share one parser; a blank
value now reads as unset. Startup check 9 reports the source, and the skip path lists every step,
which it did not before.

+++ src/cardano/treasury_datum.rs
@@ -34,7 +34,7 @@ pub struct TreasuryConfig {
     /// group x-only key.
     pub y_51: bitcoin::key::UntweakedPublicKey,
     pub y_fed: bitcoin::key::UntweakedPublicKey,
-    pub federation_csv_blocks: u32,
+    pub federation_csv_blocks: u16,
     /// The current treasury Bitcoin UTxO, tracked OFF-CHAIN by the SPO (spec
     /// §640/§1677: "known from the previous TM's change output, or from protocol
     /// bootstrap"). heimdall builds every TM, so in steady state it knows its own
